수입/지출 내역 작성 화면 DataLayer단 구성하기 완료 🔥

+++ app/src/main/java/com/seom/accountbook/model/graph/OutComeByCategory.kt
@@ -3,7 +3,7 @@ package com.seom.accountbook.model.graph
 import com.seom.accountbook.model.BaseCount
 
 data class OutComeByCategory(
-    override val id: Int, // categoryId
+    override val id: Long, // categoryId
     override val name: String, // category 이름
     override val color: Long, // category 색
     override val count: Long // category 별 지출금액
